Fix liquid hydrogen and non-integral tank calculations
Corrects the calculation of ullage and liquid volumes in append_liquid_hydrogen_tank_unknown_and_residual.py, updates residual assignments in compute_liquid_hydrogen_tank_performance.py for proper array handling, and removes unnecessary float conversion in compute_non_integral_tank_volume.py. These changes improve consistency and accuracy in fuel tank property computations.

+++ RCAIDE/Library/Methods/Powertrain/Sources/Fuel_Tanks/Non_Integral_Tank/compute_non_integral_tank_volume.py
@@ -215,7 +215,7 @@ def compute_bwb_aft_tank_volume(fuel_tank, wing):
             raise ValueError('Specified fuel mass greater than mass of fuel capable of being stored in fuel tank') 
         fuel_tank.fuel.volume_properties.net_volume = tank_volume_i
     else:
-        fuel_tank.fuel.mass_properties.mass = float(tank_volume_i *  fuel_tank.fuel.density)
+        fuel_tank.fuel.mass_properties.mass = tank_volume_i *  fuel_tank.fuel.density
         fuel_tank.fuel.volume_properties.net_volume = tank_volume_i
  
     # fuel tank origin
